requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cron-descriptor==1.4.5
cycler==0.12.1
Django==3.2.25
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-timezone-field==7.0
djangorestframework==3.12.4
drf-spectacular==0.15.1
flower==2.0.1
fonttools==4.53.1
humanize==4.10.0
importlib_resources==6.4.4
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.0.1
packaging==24.1
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyparsing==3.1.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.0
